feat: filter chips icons
